Refactor decorator for C test spec
